visibleValue was recalculated using the old props.
Fixes issue #1117.

Instead of using the `props` argument that is passed in to
`componentWillReceiveProps`, `visibleValue` was recalculated using the old
`this.props.precision`.

Renamed the argument from `props` to `newProps`, and calculated
`visibleValue` using `newProps.precision`.

Added a unit test to replicate the issue / verify the fix (which uses enzyme
to set props and check the state of the component).