Remove DraggableRow and use HOC on TableRow instead.
Remove `DraggableRow`, and instead use a HOC `extendWithDragAndDrop` to add the drag and drop functionality to the given component.

`extendWithDragAndDrop` applies the drag and drop functionality and keeps the existing functionality of the given component by inheriting from it.

In the `Table` component constructor, set the value of `this.RowComponent` based on whether `props.draggableRows` is true. If it is `true` then apply the HOC to TableRow. `this.RowComponent` is then used when rendering `tableContent`.