Add tests for draggableRows behaviour
Add test to verify that a `DroppableBody` is rendered when we create a table
with draggableRows enabled e.g. `<Table draggableRows={ true }>...</Table>`.

No props are expected for `DroppableBody`, so they have been removed and
replaced with a `data-dragdropcontext` attribute instead.

Remove the `draggable` prop from `TableRow`, and the associated tests. Set
draggable CSS class on draggable rows instead.

Instead, if a row is converted to a draggable row, add `{ className:
'carbon-table-row__draggable' }` to the existing `props`.