Add demo for drag and drop table rows.
Added action and store method to update the data.

The data itself is hard-coded for now in demo/stores/component/component.js.
This may be moved to the data directory and loaded in a similar fashion to how
we load the countries data.

The draggable-context definition has been updated to use the action as the
`moveItem` callback.