Renamed with-drag-and-drop to drag-and-drop
Add drag-and-drop js and scss files

Add drag-and-drop.js, which exports all the other modules from the sub-folders
within the drag-and-drop folder.

Update package.json to reference `drag-and-drop.js` and `drag-and-drop.scss`.

Move with-drag-and-drop.js to ./with-drag-and-drop, and remove the
with-drag-and-drop.scss file.

Move item-types.js to the ./item-types folder, and add
./item-types/package.json.

Update import paths for drag and drop components