Fix warnings on component specs
Fix warnings on required props being undefined on:

* ButtonToggle DropdownFilterAjax FormSummary Table Form

The warnings in `Table` were because the `actions` prop was undefined, which is
a required prop of the `ActionToolbar` component (which is used by the `Table`
component).

The `Form` warnings were because the `activeInput` prop was missing from the
component's `propTypes`. This has now been added.

Replace test utils shallow renderer with `react-test-renderer/shallow` instead.