Add aria attributes to sortable columns
Make table headers clickable / focusable by wrapping the contents of the header in a `<a>` tag.

Apply the `aria-sort` to the sorted `<th>` to indicate the current sort direction applied to that column.

Add the `aria-label` attribute to the sorted column's `<a>` tag to indicate that the column is sortable, the current sort direction (if applicable), and which sort will be applied when the link is activated.