Fix ESLint errors in Decimal component
Add `name`, `value`, `onBlur`, and `onKeyDown` to props, and add `onBlur` and
`onKeyDown` to the definition's `hiddenProps`.

Disable the warning about the unused align prop, as it is used by the `Input`
decorator that is applied to this component.

Changes to some props, and how `refs` are referenced, caused some tests to
fail. These have been fixed.