Apply overflow hidden to <html> when dialog open
Use the dialog `onOpening` and `onClosing` functions to add and remove
a `data-fullscreendialogopen` attribute to the `<html>` tag: the data
attribute is added when the dialog is open, and removed when the dialog
is closed.

Add the CSS to set `overflow: hidden` using the
`[data-fullscreendialogopen]` selector.

Change the `onOpening` and `onClosing` getters defined in `Modal` to
functions, as they aren't really getters.

Remove the `carbon-dialog-full-screen__dialog--open` code and CSS.