Fix focus issue with Dialog component
An issue was found in Dialog where the changes made in #1340 conflicted
with trying to set the focus to the first form field in a Dialog: it
would the focus to the form field, then change the focus to the dialog
element and trigger a validation error if the form field was a required
field.

The fix involved:

* Removing the `focusDialog` from `Dialog`, and adding props for
`onOpening`, `onClosing`, `onCloseKeyDown`, and `autoFocusCloseIcon`.

* Making `Alert` no longer inherit from `Dialog`. Instead it is composed
of a `Dialog` with the props set accordingly.

The `autoFocusCloseIcon` prop, which takes a boolean value:

* if true then the close icon receives focus when the dialog opens

* if false then the close icon does not receive focus when the dialog
opens (the develop is then free to set focus as they see fit).

A further change was made to the close icon in the `Dialog`. It renders
an `Icon` wrapped in an `<a>` tag. This fixes an issue where if the
close icon had keyboard focus, you couldn't close the dialog by pressing
the Enter key: now you can.