Fix Jest coverage for TableAjax
`handleResponse` was failing code coverage as there was no `else` path
taken.

Add a `onAjaxError` function to `TableAjax` props, and call this if an
Ajax request results in an error.

If an Ajax request results in an error and no `onAjaxError` was passed
as a prop, log a warning about the Ajax error to the console.