Update the library used by Date component
Switched from react-date-picker to react-day-picker as the old library
is no longer being maintained.

https://github.com/gpbl/react-day-picker

Note that moment js has been specifically not included in Date and is
isolated to within DateHelper. This is to abstract it incase it is ever
changed.

To test the locale switching on the Carbon demo site you can visit:
http://localhost:8095/components/date-input?locale=fr
(only French is currently supported)