Add support for aria-describedby to Table component
Add an `ariaDescribedBy` prop to the `Table` component.

This is a string, which points to the HTML id attribute of an element
containing a description of the table.

When the `Table` component renders it renders an `aria-describedby`
attribute using the value of the `ariaDescribedBy` prop.

Add ariaDescribedBy prop to the table definition.

Update release notes.