Add aria-selected and roles to Dropdown component
The li is a Non-interactive HTML element and thus requires a role to
tell screen-readers how it can be interacted with.

https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/ARIA_Tec
hniques/Using_the_listbox_role

The role of listbox for the container and option for the items identify
the element that creates a list from which a user may select one item.

https://www.w3.org/TR/wai-aria/states_and_properties#aria-selected
Aria-selected indicates the current "selected" state and should be used
in tandem with the role ‘option’.