test(jest): downgrade jest to 24.9.0
After running npm install tests were failing. After running npm ci tests
were passing. It appears this was due to multiple versions of
jest/jest-cli.

jest-cli has been removed from the dev-dependencies so we now use the
version that is bundled with jest.

The action-popover snapshots had been generated with 25.x.x and were
updated to work with 24.9.0.

It is not possible at this time to update to 25.x.x as there are missing
style objects and some focus tests are failing.