fix(icon-button): 'esc' key triggers onAction closing a modal
when a Dialog or DialogFullscreen Component has disableEscKey prop set to true
and a close icon is focused, the Modal was still being closed on ESC key pressed.

ESC key should be handled globally by a component in which the button is rendered.

BREAKING CHANGE: esc key will no longer trigger onAction prop