feat(text-editor): add support for validation icon and error styles
The `ValidationIcon` is rendered when a string value is passed to either the `error`, `warning` or
`info` prop. THe component has been slightly refactored to use outlines instead of borders to
prevent the content jumping when the styling is updated based on the validation state. When the
component is in an error state, the outline and counter text colour are updated to reflect it.