fix(numeral-date): prevent error being thrown when default date format is used
When no value is passed to the `dateFormat` prop it throws an error even though it has a default
value assigned in the top level prop destrucure of the component. This is because at the point where
the prop types are checked the value is not initialised to the default.

fix #3264