refactor(decimal): convert class to functional component
Decimal has been refactored to be a functional component from a class based component. This means
that you are no longer able to extend this component.

BREAKING CHANGE: Decimal has been refactored to be a functional component from a class based
component. This means that you are no longer able to extend this component.

fixes FE-3640