fix(multi-select): prevent re-renders from setting selectedValue to undefined
Component no longer set the selectedValue to `defaultValue` on re-renders: this was causing issues
as in some demos no `value` or `defaultValue` is passed so when the theme changes and the component
re-renders the new `selectedValue` is `undefined`. Also updates teh `isExpectedOption` function to
handle fro empty strings and objects.

fix #4482