fix(date): enforce ISO date in value prop
sanitize method of the internal date helper was converting all dots,
dashes and spaces into slashes in a passed down date string,
which triggered an error when comparing that date with available formats.
This functionality was redundant as the validation is already being done
by the moment.js.

Fixes #3945
Fixes #4453