fix(date): ensure day value is validated against format with the month value included
Ensure that any update to the day value also includes the current month value that is in the input
to prevent it using the month from today's date. Without this change valid day values will be
rejected when they are outside the today's month's valid days, eg. 29 when the today's month is
February.

fix #4853