fix(flat-table/sort): fix color of sort arrows
Make Sort component pass the color as a prop to Icon rather than just adding as a new CSS rule in
sort.style.js. This ensures that the CSS for the default icon color in flat-table rows is not
applied, so there is no fragility over CSS selector specificity.

fix #4984