feat(select): make select list respond to pageup/pagedown keys
Make it so that the option list for a Select scrolls up/down when the pageup or pagedown key is
pressed. This also means that it responds to voice commands to scroll up and scroll down, so
keyboard users and those using voice control can navigate the options much more quickly.