feat(sidebar): add aria-labelledby referencing title if not provided
Give guid as id to sidebar header container, and provide this as the aria-lablledby to the sidebar
container if no aria-label or aria-labelledby prop is provided. This improves accessibility and
makes Sidebar consistent with DialogFullScreen. It also fixes an issue with the previous commit
where, with Voiceover on, opening the sidebar places focus on the header rather than the sidebar
container.