test(simple-select): fix cypress arrow key tests to not click
The "open select list using arrow key" test previously clicked on the select text before pressing
the arrow key. This makes the test pointless as the click automatically opens the list, so the
keypress becomes irrelevant. This test passed even if the keyboard functionality was deliberately
broken - by removing the click this is no longer the case.