docs(link): add usage guidance to onClick story
Providing an onClick to the Link component is bad for accessibility as it renders an HTML button
that looks like a link. This isn't a pattern Carbon should encourage, but in lieu of removing it a
note has been added to the docs explaining that this shouldn't be used and giving more accessible
alternatives. Also the text of the component used in this story has been updated to be more
accurate.