fix(split-button): add new keyboard functionality to component
Implements `useMenuKeyboardNavigation` to update the keyboard navigation behaviour
for `SplitButton`. Up and down arrow keys can still be used for navigating children
buttons but there is no longer any looping back around. When the children container
is open tabbing can now be used to navigate the children buttons, the tab key takes
you down the list and back tab goes up. When the last child button is focused and
tab is pressed it will close the container and move focus to the next focusable
element on the page. When the shift and tab keys are pressed and the first child
button is focused it will close the container and return focus back to the main
button control. Pressing the escape key when the children container is open will
now close the list and return focus back to the main button control. All other
keyboard functionality remains the same

fix #4522