fix(submenu): make menu open on click even when clickToOpen prop missing
Currently the clicking on a menu item that has a submenu does not open the submenu unless the
clickToOpen prop is provided. Instead the menu opens on hover or when certain keys such as enter and
space are pressed. However the NVDA screen reader does not respect key handlers on buttons when in
the default browse mode, as these are used for its internal functions, and buttons unlike links and
form inputs do not trigger a change to focus mode. This is fixed by ensuring that all submenus open
on click, as space and enter on a button always trigger the click action, even with NVDA. No
problems are caused by this because it is impossible to click on a menu without also hovering.

fix #4797