fix(multiselect): ensure value prop is always used when controlled
When a controlled component has a custom onchange that sometimes does not update the value prop, the
component was not respecting this and still updating its internal value, essentially reverting the
component to being uncontrolled. This refactor, as well as fixing the bug, treats the controlled and
uncontrolled cases differently when the value changes so there should be less risk of similar bugs
in future.

fixes #5259