fix(pager): change styles so all items are exposed to accessibility tree
When using Chrome, when the Previous and Next buttons are missing due to only 1 page being present,
screenreaders such as NVDA and VoiceOver were treating "items" and "page", and the total page number
and "X items" messages, as the same element when moving through one item at a time. This appears to
be due to a bug or feature in Chrome whereby generic elements (div or span) with display property
inline or flex were not being exposed to the accessibility tree at all. So "of X" and "page" have
been changed from inline to block display (div in place of span), to fix this.

fix #4637