fix(menu): fix incorrect HTML output by ScrollableBlock
Ensure that the HTML output by a ScrollableBlock within a submenu does not incorrectly put divs
between <ul> and <li> elements - this invalid markup is a WCAG failure in itself and also causes the
menu to be announced incorrectly by some browser/screenreader combinations. Scrollable sections are
now rendered as child lists, and enabled by using the new `scrollable` prop on a MenuItem (either
the top-level item that opens the submenu, or an item within a submenu), with the ScrollableBlock
component now being a purely internal component.

BREAKING CHANGE: ScrollableBlock is now no longer usable. To get the same functionality, add the
`scrollable` prop to the MenuItem that contains it - along with the `scrollHeight` and
`scrollVariant` props that replace the previous `height` and `variant` props of ScrollableBlock. For
cases where ScrollableBlock would have wrapped a subset of the menu items in a submenu, the previous
menu item is now treated as a parent and should wrap the following MenuItems, as well as its text,
and be given the new props. Note that cases where a ScrollableBlock wrapped a subset of menu items
including the first one are no longer supported.