fix(select): ensure aria-activedescendant is always valid
make the Select List always part of the DOM, and hide it with CSS instead. This allows the component
to ensure that the guids provided as IDs to the Option children are always stable, so that the
aria-activedescendant attribute of the Select Textbox always refers to the ID of an existing
element. Also make the aria-controls attribute of the textbox always present so that screenreaders
can know the list of options available when focused on the textbox.

Adjust FilterableSelect component to ensure that when controlled the handleGlobalClick handler
always reads the updated value, to avoid a race condition when adding a new option which caused
a cypress test failure.

fix #5289

refactor(popover): replace TempModalContext with ModalContextProps

test(select): check visibility of options directly rather than isOpen prop

refactor(select-list): move isOpen check inside event handler