fix(popover-container): ensure clicks inside popover do not close it
Fix bug where portals inside a component did not automatically work with useClickAwayListener.
Refactor useClickAwayListener to no longer take an array of refs and return a handler which can be
attached to the parent component and stops the clickAway handler firing on the document - this
automatically handlers any React children no matter whether they are rendered as DOM children or
not. Remove any refs in parent components which are now unneeded.

fix 5322