fix(tabs, icon): override behaviour of icon with tooltip when rendered inside tab title
When an `Icon` with a `Tooltip` is rendered within a `TabTitle`, the show/ hide functionality is
overridden so that the `Tooltip` displays when the `TabTitle` has focus or the mouse hovers over and
is hidden when loses focus or the mouse leaves. The tabIndex for the `Icon` is also set to undefined
so that it cannot be tabbed to.

fix #4663