fix(dialog, sidebar): ensure aria-modal is only added to top modal
Currently we add aria-modal=true to the container div of each modal component (dialog,
dialog-full-screen, sidebar). This ensures screenreaders do not announce content outside the modal.
However when opening one modal from another, VoiceOver on Safari refuses to read the content on the
new modal while the old one still has aria-modal=true. The workaround is to ensure aria-modal=true
only appears on whatever the current top modal is. This requires keeping track of the current top
modal in a context which has been added to CarbonProvider, as well as keeping a list of setter
functions in the window object to ensure that multiple Carbon Providers in the same application
(potentially even from different versions of Carbon) still update each other's context correctly.

fix #5303