fix(multi-select): ensure aria-selected has the correct value
Currently on all selects, including MultiSelect, we set aria-selected=true on only the option that
is currently focused. In particular, when using a screenreader and navigating through the options,
"selected" is announced for each one. This may be correct for single select (because navigating
through an item puts its text value into the select textbox), it is definitely not correct for
MultiSelect and means the user has no way of knowing which values are currently selected when using
a screenreader.

fix #5272