docs(link): add hrefs to Menuitems on isSkipLink story
Ensure that the skip link is demonstrated properly by making the menu items below them tabbable, so
that the advantage of the skiplink in skipping over them can be shown. Previously they were not
tabbable due to lacking an href and therefore being rendered as HTML anchor tags without an href,
which are not tabbable.