fix(tags): update TagProps interface so only the data tag types are exported
Updates the `TagProps` interface to ensure only the `data-` props are exported and extended in the
public interfaces. A new private `RestProps` interface has been added to support passing the rest
props object into the `tags` util. The previous implementation was preventing `intellisense` from
displaying the correct types for the `Sidebar` component; instead it was being assigned `any`.

fix #5534