fix(form): fix form styling when inside a sidebar
A form with sticky footer previously hard-coded the margin and width of the footer to match the
default padding of a sidebar. This resulted in incorrect styling when custom padding is used for the
sidebar as is now allowed. To fix this the width and margin of the form footer now depend on the
padding applied to the sidebar.

fix #5404