fix(select): ensure IDs of select options remain stable on prop changes
When a prop is updated to one of the Select components - including when the value is updated in a
controlled select - the guid IDs of each option child are regenerated. Amongst other problems, this
causes an Axe error because the aria-activedescendant value is no longer valid. Instead the IDs
should only update when a new child is added or removed.