fix(pager): ensure there is no margin-bottom on page number input
When rendered inside a Form component, the numberic input on Pager gets CSS rules to given it
margin-bottom equal to the Form's field-spacing prop. In order to ensure the sub-components of Pager
stay in line it always needs 0 margin-bottom, so a CSS rule has been added to set this, and made
more specific that the one in Form.

fix #5650