feat(multi-action-button): surface width prop to allow width to be set for component
adds width prop which allows consumers to set a desired width for `MultiActionButton`
a viable alternative to the `size` prop. Default styling behaviour has also been changed on
`Button` components whilst children of `MultiActionButton` to justify-content: space-between
when prop is passed. Which ensures text and icon of `Button` appear on the opposite side.

fix #5547, fix #5672