fix(menu, submenu): refactor components to remove children iteration and cloning
Refactors the `Menu`, `Submenu`, `MenuItem` and `ScrollableBlock` to remove any
code that relies on children composition. By iterating over the children and/ or
relying on cloning to pass additional props we are limiting the ability of
consumers to wrap the child components. Code that relied on that approach has
been removed in favour of context and querySelectors to achieve the same
functionality.

fix #5592