feat(anchor-navigation): changed behaviour to improve accessibility
Make the links in Anchor Navigation separate tab stops, and remove the arrow-key navigation it
previously had. Make the links only work with Enter rather than Space. When activating a link,
ensure that focus goes to the container of the corresponding section (without visible focus
indicator), so that subsequent tabs navigate through that area but without breaking the user
experience when that section has no focusable elements.

fix #5283