fix(focus-trap): ensure focus trap behaviour does not work except in the top modal
Nested dialogs could give rise to a situation where if focus is lost (eg by clicking a non-focusable
element), when pressing tab the first element of the wrong focus trap is focused, thereby trapping
focus in the wrong container. This is now fixed by ensuring that if there are multiple modals open
the focus trap does nothing except in the top modal.

fix #5754