fix(link): change display to inline so that text can wrap
Link components currently use `display: inline-block` for both the anchor element itself and one of
the containing components. This means that when the text of a link is long enough to wrap onto a new
line, the entire link element is moved to the next line, creating unnecessary blank space. This PR
fixes that by removing the `display: inline-block`, except when used in a MenuItem - that component
will never be used inline and the inline-block display is needed for the background color on hover
to fill the entire vertical space.

fix #5774