fix(multi-action-button, split-button): open menu on click
The additional buttons are now shown on click of the toggle button, rather than just on hover. This
means that Voice Control users can activate the button by just saying its name (or the appropriate
number if going through "show numbers"). Also tidy up by removing aria-haspopup on the surrouding
div, this was already on the toggle button itself and from ARIA 1.2 is no longer valid on generic
elements like a div.

fix #5615