refactor(textbox): remove labelId prop
labelId prop was always intended to be an internal prop,
to be used to set the label id generated in the select,
and provide description to be read by screen readers.

Label id could be generated using common pattern,
so there is no need to pass the prop down.
The aria-labelledby value could be passed down directly to the input instead.

BREAKING CHANGE: labelId prop is removed