test(accordion, flat-table): add helper for testing rotation angle
Currently the cypress tests for Accordion and FlatTable fail when run locally, although they pass on
Github CI. This is because recent versions of Chrome (since v109) appear to calculate CSS transform
matrices more accurately for rotations - the Github CI is currently running on v107, so to get this
to pass we've had to test against specific CSS values that are not quite zero. This not only causes
failure when using updated versions of Chrome, it's a very fragile way to test in general. To avoid
this we should refactor to use a test utility specifically designed to compute the rotation angle
from such a CSS matrix string, while accounting for rounding errors.