feat(button-toggle, button-toggle-group): update HTML for accessibility
Rewrite the ButtonToggle and ButtonToggleGroup components so that they render as a set of HTML
button elements, rather than an HTML radio button group that visually looks like a group of buttons.

BREAKING CHANGE: Changes to props in the ButtonToggle component:
- `defaultChecked`, `guid` and `onChange` have been removed
- `onClick`, `onBlur` and `onFocus` have new types
Changes to props in the ButtonToggleGroup component:
- all validation-related props (`validationOnLabel`, `error`, `warning` and `info`) have been
removed
- `onBlur` has also been removed (the main purpose of this will have been to perform validation)
- the `onChange` prop has a new type. The event argument now has an HTML `button` as its target,
with the `value` and `name` props of the corresponding `ButtonToggle` component passed as optional
second and third arguments
Other important changes to be aware of:
- import paths changed: both components are now named exports from `components/button-toggle`
- these components can no longer be used in an "uncontrolled" fashion, the state of which button
(if any) has been selected must be managed by consumers using React
- the `name` prop on ButtonToggleGroup is no longer mandatory. Both this and the corresponding
prop on ButtonToggle now provide nothing except an internal label which can be used in an
`onChange` handler. These have been kept for backwards compatibility and will soon be removed
- the `checked` prop on ButtonToggle has likewise been kept only temporarily for backwards
compatibility, and is an alias for the new `pressed` prop which sets the `aria-pressed`
attribute on the underlying HTML button
- a new `allowDeselect` prop has been added to `ButtonToggleGroup` to allow a currently-
selected button to be deselected, leaving no selection
- an `aria-label` prop can be passed to `ButtonToggleGroup`, and this *must* be used for
accessibility if the visual `label` is omitted