refactor(tile-select): convert component to typescript
This commit refactors the TileSelect component to use
TypeScript. The name prop is now required for
TileSelectGroup, and the onChange handler in both TileSelect
and TileSelectGroup components has been updated to
properly type the arguments passed to it. This change was
necessary because the custom event passed to the onChange
handler when a tile is deselected was not previously handled correctly.

BREAKING CHANGE: The name prop is now required for
TileSelect. Additionally, the onChange handler in TileSelect and
TileSelectGroup components has been updated to properly
type arguments, including custom events. Also inputRef prop has been removed.