feat(help): improve accessibility
Remove `role="button"` from help icon, and remove `aria-label` and `aria-describedby`. The relevant
text is now included in the DOM directly in styles that make it not visible but are picked up by
screenreaders - ensuring a good experience for screenreader users. Deprecate the `tooltipId` prop
which is no longer necessary for accessibility, and deprecate `aria-label` and replace with
`accessibilityLabel` (as the prop's value is no longer used for an `aria-label` attribute)

fix #5569