fix(select): stop select list triggering refocus if closed inside a modal
The select list is currently treated as a modal, which means that when opened from another modal and
then closed, focus is placed where it previously was before the select list was opened - which is
the select textbox itself. This is unwanted behaviour if closing the list by clicking outside -
particularly if the `openOnFocus` prop is set on the select, in which case the select never actually
gets closed.

fix #5898