fix(link): ensure button has a type when used
Link component was used as a button currenly does not supply the button with a type. This has caused
a bug when `Pager` is used in a `Form`. The bug was causing the `onPrevious` callback function in
`Pager` to be called in an onClick event when the enter key pressed.

fixes #5676