fix(menu-fullscreen): ensure correct tab order when Search with button rendered as a child
Ensures that the correct tab order is maintained when a `MenuItem` with
a `Search` input child with `searchButton` enabled but no value. Currently the
button disappears when the input is blurred and there is no value meaning focus
is placed on the document body instead. This update ensures that the next
item in the Menu is focused instead.

fix #5959