fix(focus-trap): prevent scroll to bottom when tabbing through
The addition of invisible "tab guard" divs at the top and bottom of each focus trap causes the
browser to scroll these into view when tabbing into them - causing the user to see the background of
the page jump to the bottom when tabbing from the bottom element in the trap to the top one, and
vice versa. This is now fixed by giving these divs a CSS `position: fixed` property, which keeps
them permanently in the viewport so stops the browser scrolling anything.

fix #5986