fix(select): reset height and width after closing
The useFloating hook, which is used by Select to position the list, correctly resets various CSS
properties to their previous values when the component is no longer open. But it failed to do this
with width or height. Since the select list computes its width when opened, this lack of cleanup
results in a hard-coded CSS width in the style attribute, which messes up the layout if any later
resizing happens. This is now fixed by cleaning up the width and height properties on close.

fix #5981