fix(menu): ensure onSubmenuOpen only fires once when submenu opens
The Submenu component was calling the onSubmenuOpen prop function on each relevant click, mouseover
or keyDown event. This results in the function being called several times on mouseover, and twice
when pressing Enter or Space (as these also cause the browser to fire a click event on the button).
This is fixed by using useEffect to only call the function once whenever the submenu's state changes
from closed to open, no matter what the cause.