fix(select): fix accessible name
Simple select and Filterable select both use a 2-part aria-labelledby that references both the
field's label and a hidden element containing the currently-selected  value. This results in the
accessible name being different depending on the selected value, which is unexpected to users who
will expect that the accessible name is simply the label.

Note that MultiSelect has been unchanged for now, pending further investigation with
UX/accessibility of the best way to make this component accessible.

As a result of these changes, Select and FilterableSelect do not automatically get an accessible
name provided, so all uses need either a `label`, `aria-label` or `aria-labelledby` to pass
automatic accessibility checks. In particular, `pager` has had to be updated to ensure the
page size select always has an accessible name even if `showPageSizeLabelBefore` is false.

fix #5945