feat(action-popover): align component with design system
aligns component with the sage design system, to ensure consumers can use the component with
expected styling and functionality. Nine different styling variations have been added
which change the alignment and padding of each child to ensure designs are correct. Also,
support for submenu's opening on the right has been included with the addition of the
`submenuPosition` prop. When set to "right" the menu-item chevron will appear on the right
of the action popover menu, and any submenu's will open from the right-hand side. Also,
when the `submenuPosition` prop is "left" or "right", the `submenuPosition` will be temporarily
set to the opposite value if the window on the left or right is deemed to be too small.
Finally, icon's are now supported in submenu's and can be included by passing the `icon`
prop to the submenu's menu item.

fix #5801